// src/app/models/crup.ts
// export class Crup {
//   constructor(
//     public create_at: string ,
//     public crup: string ,
//     public uid?: string ,
//   ){}

// }


// export interface Crup {
//   crup: string;
//   create_at: string;
//   uid?: string;
// }

export class Crup {
  create_at?: number;
  crup?: string;
  uid?: string;

}